# Optimized some container equality functions