--- a/include/roaring/containers/run.h
+++ b/include/roaring/containers/run.h
@@ -14,6 +14,7 @@
 #include <roaring/containers/perfparameters.h>
 #include <roaring/portability.h>
 #include <roaring/roaring_types.h>
+#include <roaring/array_util.h>
 
 /* struct rle16_s - run length pair
  *
@@ -522,8 +523,14 @@ static inline int32_t run_container_size_in_bytes(
 /**
  * Return true if the two containers have the same content.
  */
-bool run_container_equals(const run_container_t *container1,
-                          const run_container_t *container2);
+static inline bool run_container_equals(const run_container_t *container1,
+                          const run_container_t *container2) {
+    if (container1->n_runs != container2->n_runs) {
+        return false;
+    }
+    return memequals(container1->runs, container2->runs,
+                     container1->n_runs * sizeof(rle16_t));
+}
 
 /**
 * Return true if container1 is a subset of container2.