Bitset/bitset, array/array and run/run container
equality functions now work by comparing raw memory.
When AVX=on, use manually crafted function.
When AVX=off, fallback to libc memcmp().

Minor improvement in run/bitset equality function:
use bitset_contains_range() instead of calling
bitset_contains() for every rle element.

Added container equality benchmark.