@andreigudkov 2018-11-10
Benchmark results.

Before optimizations, AVX=off:
```
[Size:   64] [NRuns:   -1] [NRuns:   -1] array_container_equals(c1, c2):  146.00 cycles per operation
[Size: 4096] [NRuns:   -1] [NRuns:   -1] array_container_equals(c1, c2):  8256.00 cycles per operation
[Size: 8192] [NRuns:   -1] [NRuns:   -1] array_container_equals(c1, c2):  16479.00 cycles per operation
[Size:65535] [NRuns:   -1] [NRuns:   -1] bitset_container_equals(c1, c2):  1422.00 cycles per operation
[Size:65536] [NRuns:   -1] [NRuns:   -1] bitset_container_equals(c1, c2):  1428.00 cycles per operation
[Size: 2048] [NRuns: 1964] [NRuns: 1964] run_container_equals(c1, c2):  3986.00 cycles per operation
[Size: 4096] [NRuns: 3875] [NRuns: 3875] run_container_equals(c1, c2):  7894.00 cycles per operation
[Size: 4096] [NRuns: 3822] [NRuns:   -1] run_container_equals_array(c1, c2):  16812.00 cycles per operation
[Size: 4096] [NRuns:   -1] [NRuns:   -1] array_container_equal_bitset(c1, c2):  20790.00 cycles per operation
[Size: 4096] [NRuns: 3830] [NRuns:   -1] run_container_equals_bitset(c1, c2):  20381.00 cycles per operation
```

Optimized version, AVX=off:

```
[Size:   64] [NRuns:   -1] [NRuns:   -1] array_container_equals(c1, c2):  57.00 cycles per operation
[Size: 4096] [NRuns:   -1] [NRuns:   -1] array_container_equals(c1, c2):  640.00 cycles per operation
[Size: 8192] [NRuns:   -1] [NRuns:   -1] array_container_equals(c1, c2):  1310.00 cycles per operation
[Size:65535] [NRuns:   -1] [NRuns:   -1] bitset_container_equals(c1, c2):  641.00 cycles per operation
[Size:65536] [NRuns:   -1] [NRuns:   -1] bitset_container_equals(c1, c2):  35.00 cycles per operation
[Size: 2048] [NRuns: 1964] [NRuns: 1964] run_container_equals(c1, c2):  613.00 cycles per operation
[Size: 4096] [NRuns: 3875] [NRuns: 3875] run_container_equals(c1, c2):  1190.00 cycles per operation
[Size: 4096] [NRuns: 3822] [NRuns:   -1] run_container_equals_array(c1, c2):  16771.00 cycles per operation
[Size: 4096] [NRuns:   -1] [NRuns:   -1] array_container_equal_bitset(c1, c2):  22541.00 cycles per operation
[Size: 4096] [NRuns: 3830] [NRuns:   -1] run_container_equals_bitset(c1, c2):  19261.00 cycles per operation
```

Optimized version, AVX=on:

```
[Size:   64] [NRuns:   -1] [NRuns:   -1] array_container_equals(c1, c2):  51.00 cycles per operation
[Size: 4096] [NRuns:   -1] [NRuns:   -1] array_container_equals(c1, c2):  532.00 cycles per operation
[Size: 8192] [NRuns:   -1] [NRuns:   -1] array_container_equals(c1, c2):  1072.00 cycles per operation
[Size:65535] [NRuns:   -1] [NRuns:   -1] bitset_container_equals(c1, c2):  560.00 cycles per operation
[Size:65536] [NRuns:   -1] [NRuns:   -1] bitset_container_equals(c1, c2):  33.00 cycles per operation
[Size: 2048] [NRuns: 1964] [NRuns: 1964] run_container_equals(c1, c2):  511.00 cycles per operation
[Size: 4096] [NRuns: 3875] [NRuns: 3875] run_container_equals(c1, c2):  1033.00 cycles per operation
[Size: 4096] [NRuns: 3822] [NRuns:   -1] run_container_equals_array(c1, c2):  13614.00 cycles per operation
[Size: 4096] [NRuns:   -1] [NRuns:   -1] array_container_equal_bitset(c1, c2):  19727.00 cycles per operation
[Size: 4096] [NRuns: 3830] [NRuns:   -1] run_container_equals_bitset(c1, c2):  18101.00 cycles per operation
```